service.requirements.txt
annotated-types==0.5.0
anyio==3.7.0
APScheduler==3.6.3
async-generator==1.10
attrs==23.1.0
Babel==2.12.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
calendar-view==2.3.2
cbapi==1.7.10
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==37.0.4
decorator==5.1.1
dnspython==2.4.2
docxcompose==1.4.0
docxtpl==0.16.7
et-xmlfile==1.1.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.3
opencv-python==3.4.13.47
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pika==1.3.2
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.15.1
pymongo==4.5.0
pyotp==2.9.0
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-telegram-bot==13.13
pytz==2023.3
PyYAML==6.0
requests==2.30.0
schedule==1.2.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
solrq==1.1.2
sortedcontainers==2.4.0
telebot==0.0.5
tornado==6.3.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
webdriver-manager==3.8.6
wsproto==1.2.0